[oh-tab-zhwi.3.1] Extract App context and tool/skill domain hooks
